Changed doubles to ints for nicer calculations.

+++ src/main/java/CritterControl/Accessories/Accessory.java
@@ -1,4 +1,4 @@
 package CritterControl.Accessories;
 
-public record Accessory(String name, double healthBoost, double strengthBoost, double speedBoost, double magicBoost) {
+public record Accessory(String name, int healthBoost, int strengthBoost, int speedBoost, int magicBoost) {
 }
